+++ extensions/womir_circuit/src/load_sign_extend/execution.rs
@@ -71,9 +71,9 @@ where
 #[repr(C)]
 struct LoadSignExtendPreCompute {
     imm_extended: u32,
-    a: u8,
-    b: u8,
-    e: u8,
+    a: u32,
+    b: u32,
+    e: u32,
 }
 
 impl<A, const LIMB_BITS: usize> LoadSignExtendExecutor<A, { RV32_REGISTER_NUM_LIMBS }, LIMB_BITS> {
@@ -115,9 +115,9 @@ impl<A, const LIMB_BITS: usize> LoadSignExtendExecutor<A, { RV32_REGISTER_NUM_LI
 
         *data = LoadSignExtendPreCompute {
             imm_extended,
-            a: a.as_canonical_u32() as u8,
-            b: b.as_canonical_u32() as u8,
-            e: e_u32 as u8,
+            a: a.as_canonical_u32(),
+            b: b.as_canonical_u32(),
+            e: e_u32,
         };
         let enabled = !f.is_zero();
         Ok((local_opcode == LOADB, enabled))
@@ -196,7 +196,7 @@ unsafe fn execute_e12_impl<
     let pc = exec_state.pc();
     let fp = exec_state.memory.fp::<F>();
     let rs1_bytes: [u8; RV32_REGISTER_NUM_LIMBS] =
-        exec_state.vm_read(RV32_REGISTER_AS, fp + pre_compute.b as u32);
+        exec_state.vm_read(RV32_REGISTER_AS, fp + pre_compute.b);
     let rs1_val = u32::from_le_bytes(rs1_bytes);
     let ptr_val = rs1_val.wrapping_add(pre_compute.imm_extended);
     // sign_extend([r32{c,g}(b):2]_e)`
@@ -204,8 +204,7 @@ unsafe fn execute_e12_impl<
     let shift_amount = ptr_val % 4;
     let ptr_val = ptr_val - shift_amount; // aligned ptr
 
-    let read_data: [u8; RV32_REGISTER_NUM_LIMBS] =
-        exec_state.vm_read(pre_compute.e as u32, ptr_val);
+    let read_data: [u8; RV32_REGISTER_NUM_LIMBS] = exec_state.vm_read(pre_compute.e, ptr_val);
 
     let write_data = if IS_LOADB {
         let byte = read_data[shift_amount as usize];
@@ -224,7 +223,7 @@ unsafe fn execute_e12_impl<
     };
 
     if ENABLED {
-        exec_state.vm_write(RV32_REGISTER_AS, fp + pre_compute.a as u32, &write_data);
+        exec_state.vm_write(RV32_REGISTER_AS, fp + pre_compute.a, &write_data);
     }
 
     exec_state.set_pc(pc.wrapping_add(DEFAULT_PC_STEP));
